Build IndieAuth routes with <blog_name>
This will let us know what blog we're working with during IndieAuth
calls.

Will allow us to implement better security so that grants and bearer
tokens issued for one blog cannot be used for another.

Will also let us work with Micropub clients which do not send a "me"
URI, e.g. Quill. Apparently this is optional; micropub.rocks does send
it, but Quill doesn't. We have to pass the "me" back in the response,
but without a separate endpoint per blog, if the client doesn't send it,
we can't know which blog we need to return the URI of.

* Remove owner_profile from configuration. This is not needed now that
  we have "me" everywhere.
* Use blog's baseuri as "me" everywhere
* Improve the appconfig .blog() method
* Add blog_name requirement to IndieAuth routes
* Move some code into indieauth bp's util module that should have
  happened in previous commit
* Renamed the example blog to 'example-blog' for more clarity